requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
Async-OAuthlib==0.0.9
async-timeout==4.0.2
attrs==23.1.0
blinker==1.5
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
discord.py==2.3.1
dnspython==2.3.0
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
oauthlib==3.2.2
priority==2.0.0
PyJWT==2.7.0
pymongo==4.4.0
python-dotenv==1.0.0
quart==0.18.4
Quart-Discord==2.1.4
redis==4.6.0
requests==2.31.0
toml==0.10.2
urllib3==2.0.3
Werkzeug==2.3.6
wsproto==1.2.0
yarl==1.9.2
